rearrange untranslated unicode tests to apptest
